requirements.txt
python-dotenv>=1.0.1
requests>=2.32.3
azure-functions>=1.22.1
azure-data-tables>=12.7.0
fastapi>=0.115.0
uvicorn>=0.30.6
jinja2>=3.1.4
numpy>=2.1.0
faiss-cpu>=1.8.0; python_version < "3.13"
